Fully over to custom bubble types
Added several new custom bubbles and specific images for all of them to differentiate them on the board. Also added a number of new effects and means for triggering them (on fire, on collide, etc) plus a new modifier (immovable). Card reward screen improved by the addition of a generalized multi-layered image support in the ImageDisplayWidget such that we can build up a display of the bubble in the preview that is the same as what it has on the board.